Delete encrypted backup after uploading to s3
solves #2
This commit exposes a new setting 'keep_encrypted_backup_after_upload'.
When set to True, the encrpted gpg backup is kept on-disk after
uploading to s3. When set to False (or not specified), the encrypted
gpg backup is deleted after being uploaded to s3.

I tested this by running `sudo python3 serverbackup.py` in the following
scenarios:
- flag not in config file: gpg deleted as expected
- flag in config file set to true: gpg not deleted as expected
- flag in config file set to false: gpg deleted as expected